hacker-news-bot/base/pyproject.toml
[tool.poetry]
name = "base"
version = "0.1.0"
description = ""
authors = ["Xiaojing Wang <2002434+wangxj03@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
absl-py = "^2.1.0"
duckdb = "^1.1.2"
fastapi = "^0.115.4"
langfuse = "^2.53.3"
numpy = "^2.1.3"
openai = "^1.53.0"
pandas = "^2.2.3"
pydantic = "^2.9.2"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
